frontend: clean up unsafe stall logic
