+++ coreblocks/frontend/frontend.py
@@ -1,8 +1,11 @@
+from typing import Sequence
+
 from amaranth import *
 
 from transactron.core import *
 from transactron.lib import BasicFifo, Connect, Pipe
-from transactron.utils import assign
+from transactron.utils import assign, logging, popcount, MethodStruct, OneHotMux
+from transactron.utils.assign import AssignArg
 from transactron.utils.dependencies import DependencyContext
 
 from coreblocks.interface.layouts import ExceptionInformationRegisterLayouts
@@ -15,8 +18,10 @@ from coreblocks.frontend.bpu.bpu import BranchPredictionUnit
 from coreblocks.cache.icache import ICache, ICacheBypass
 from coreblocks.cache.refiller import SimpleCommonBusCacheRefiller
 from coreblocks.interface.layouts import *
-from coreblocks.interface.keys import FlushICacheKey, RollbackKey
+from coreblocks.interface.keys import FlushICacheKey, RollbackKey, UnsafeInstructionResolvedKey
 from coreblocks.peripherals.bus_adapter import BusMasterInterface
+
+log = logging.HardwareLogger("frontend")
 
 
 class RollbackTagger(Elaboratable):
@@ -116,6 +121,9 @@ class CoreFrontend(Elaboratable):
         self.consume_instr = Method(o=self.gen_params.get(SchedulerLayouts).scheduler_in)
         self.redirect = Method(i=layouts.backend_redirect)
 
+        self._resume_from_unsafe = Method(i=layouts.backend_redirect)
+        self.connections.add_dependency(UnsafeInstructionResolvedKey(), self._resume_from_unsafe)
+
         self.get_exception_information = Method(o=gen_params.get(ExceptionInformationRegisterLayouts).get)
 
     def elaborate(self, platform):
@@ -130,7 +138,6 @@ class CoreFrontend(Elaboratable):
 
         m.submodules.fetch = self.fetch
         self.fetch.cont.provide(self.instr_buffer.write)
-        self.fetch.stall_unsafe.provide(self.stall_ctrl.stall_unsafe)
         m.submodules.instr_buffer = self.instr_buffer
 
         m.submodules.ftq = self.ftq
@@ -161,7 +168,6 @@ class CoreFrontend(Elaboratable):
         m.submodules.output_pipe = self.output_pipe
 
         m.submodules.stall_ctrl = self.stall_ctrl
-        self.stall_ctrl.redirect_frontend.provide(self.redirect)
         self.stall_ctrl.get_exception_information.provide(self.get_exception_information)
 
         rollback = Method(i=self.gen_params.get(RATLayouts).rollback_in)
@@ -171,10 +177,20 @@ class CoreFrontend(Elaboratable):
         def _(ftq_ptr, pc):
             flush(m)
             self.ftq.backend_redirect(m, ftq_ptr=ftq_ptr, pc=pc)
-            self.stall_ctrl.on_redirect_frontend(m)
 
         @def_method(m, rollback)
         def _(tag, pc, ftq_ptr):
+            self.redirect(m, ftq_ptr=ftq_ptr, pc=pc)
+
+        def resume_combiner(m: Module, args: Sequence[MethodStruct], runs: Value) -> AssignArg:
+            # Make sure that there is at most one caller - there can be only one unsafe instruction
+            log.assertion(m, popcount(runs) <= 1)
+            return OneHotMux.create(m, [(runs[i], args[i]) for i in range(len(args))])
+
+        @def_method(m, self._resume_from_unsafe, nonexclusive=True, combiner=resume_combiner)
+        def _(ftq_ptr, pc):
+            log.info(m, True, "Unsafe instruction resolved to pc=0x{:x}", pc)
+
             self.redirect(m, ftq_ptr=ftq_ptr, pc=pc)
 
         @def_method(m, flush, nonexclusive=True)
